@@ -1,17 +1,20 @@
-#include "Shared/Filesystem.hpp"
+#include "Crunch/Filesystem.hpp"
 
 #include <fstream>
 #include <iterator>
+#include <cstring>
 
 namespace Cr
 {
 
-std::vector<u8> read_binary_file(const std::string& path)
+std::vector<U8> read_binary_file(const std::filesystem::path& path)
 {
     std::ifstream file;
 
     file.exceptions(std::ifstream::badbit);
     file.open(path, std::ifstream::binary);
+
+    CR_ASSERT_THROW(file.is_open(), "Failed to open {}: {}", path.string(), std::strerror(errno));
 
     return {std::istreambuf_iterator<char>(file), {}};
 }
